test/requirements_lnx_cu12.txt
--extra-index-url https://download.pytorch.org/whl/cu124

numpy==1.26.4; python_version < '3.13'
numpy==2.3.4; python_version >= '3.13'
cupy-cuda12x==13.6.0; python_version <= '3.13'
pillow==11.2.1; python_version < '3.13'
pillow==12.0.0; python_version >= '3.13'
torch==2.6.0+cu124; python_version < '3.13' and platform_machine == "x86_64"
opencv-python==4.11.0.86
scikit-image==0.24.0; python_version < '3.10'
scikit-image==0.25.0; python_version >= '3.10' and python_version < '3.14'
scipy==1.13.1; python_version < '3.10'
scipy==1.15.1; python_version >= '3.10' and python_version < '3.14'
pytest==8.3.4
iniconfig<2.1.0; python_version < '3.10'
pynvml==11.5.3
--extra-index-url https://pypi.nvidia.com
nvidia-nvjpeg-cu12
nvidia-nvjpeg2k-cu12
nvidia-nvtiff-cu12
nvidia-nvcomp-cu12
nvidia-cuda-runtime-cu12
